GraphQL and Apollo GraphQL for Engineers
Learn GraphQL, and how to build and administer Apollo GraphQL server or middleware to create a fully functional, cloud based or on premise persistent data solution for web or mobile applications, or to mirror data between MongoDB, Redis, CouchDB, or MySQL services. - More will be added overtime.

This also will include methods to implement database and backend scalability for the corporate long-term, from a Database Administration, DevOps/Platform Engineering, and a Developers perspective.